LEP-6584
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@17634 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/service/persistence/OrganizationPersistenceImpl.java
@@ -24,6 +24,11 @@ package com.liferay.portal.service.persistence;
 
 import com.liferay.portal.NoSuchOrganizationException;
 import com.liferay.portal.SystemException;
+import com.liferay.portal.kernel.dao.jdbc.MappingSqlQuery;
+import com.liferay.portal.kernel.dao.jdbc.MappingSqlQueryFactoryUtil;
+import com.liferay.portal.kernel.dao.jdbc.RowMapper;
+import com.liferay.portal.kernel.dao.jdbc.SqlUpdate;
+import com.liferay.portal.kernel.dao.jdbc.SqlUpdateFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
 import com.liferay.portal.kernel.dao.orm.FinderCacheUtil;
 import com.liferay.portal.kernel.dao.orm.Query;
@@ -46,14 +51,6 @@ import com.liferay.portal.service.persistence.impl.BasePersistenceImpl;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
-import org.springframework.dao.DataAccessException;
-
-import org.springframework.jdbc.core.SqlParameter;
-import org.springframework.jdbc.object.MappingSqlQuery;
-import org.springframework.jdbc.object.SqlUpdate;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.sql.Types;
 
 import java.util.ArrayList;
@@ -1176,7 +1173,7 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		}
 	}
 
-	public List<Organization> findWithDynamicQuery(DynamicQuery dynamicQuery)
+	public List<Object> findWithDynamicQuery(DynamicQuery dynamicQuery)
 		throws SystemException {
 		Session session = null;
 
@@ -1195,16 +1192,16 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		}
 	}
 
-	public List<Organization> findWithDynamicQuery(DynamicQuery dynamicQuery,
+	public List<Object> findWithDynamicQuery(DynamicQuery dynamicQuery,
 		int start, int end) throws SystemException {
 		Session session = null;
 
 		try {
 			session = openSession();
 
-			dynamicQuery.compile(session);
-
 			dynamicQuery.setLimit(start, end);
+
+			dynamicQuery.compile(session);
 
 			return dynamicQuery.list();
 		}
@@ -1745,7 +1742,7 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				return list;
 			}
 			catch (Exception e) {
-				throw new SystemException(e);
+				throw processException(e);
 			}
 			finally {
 				closeSession(session);
@@ -1816,7 +1813,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		}
 	}
 
-	public boolean containsGroup(long pk, long groupPK) {
+	public boolean containsGroup(long pk, long groupPK)
+		throws SystemException {
 		boolean finderClassNameCacheEnabled = OrganizationModelImpl.CACHE_ENABLED_GROUPS_ORGS;
 
 		String finderClassName = "Groups_Orgs";
@@ -1837,13 +1835,19 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		}
 
 		if (result == null) {
-			Boolean value = Boolean.valueOf(containsGroup.contains(pk, groupPK));
+			try {
+				Boolean value = Boolean.valueOf(containsGroup.contains(pk,
+							groupPK));
 
-			FinderCacheUtil.putResult(finderClassNameCacheEnabled,
-				finderClassName, finderMethodName, finderParams, finderArgs,
-				value);
+				FinderCacheUtil.putResult(finderClassNameCacheEnabled,
+					finderClassName, finderMethodName, finderParams,
+					finderArgs, value);
 
-			return value.booleanValue();
+				return value.booleanValue();
+			}
+			catch (Exception e) {
+				throw processException(e);
+			}
 		}
 		else {
 			return ((Boolean)result).booleanValue();
@@ -1863,8 +1867,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			addGroup.add(pk, groupPK);
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1876,8 +1880,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			addGroup.add(pk, group.getPrimaryKey());
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1890,8 +1894,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addGroup.add(pk, groupPK);
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1905,8 +1909,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addGroup.add(pk, group.getPrimaryKey());
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1917,8 +1921,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			clearGroups.clear(pk);
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1929,8 +1933,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			removeGroup.remove(pk, groupPK);
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1942,8 +1946,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			removeGroup.remove(pk, group.getPrimaryKey());
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1957,8 +1961,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				removeGroup.remove(pk, groupPK);
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1972,8 +1976,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				removeGroup.remove(pk, group.getPrimaryKey());
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -1988,8 +1992,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addGroup.add(pk, groupPK);
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -2005,8 +2009,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addGroup.add(pk, group.getPrimaryKey());
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Groups_Orgs");
@@ -2082,7 +2086,7 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				return list;
 			}
 			catch (Exception e) {
-				throw new SystemException(e);
+				throw processException(e);
 			}
 			finally {
 				closeSession(session);
@@ -2153,7 +2157,7 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		}
 	}
 
-	public boolean containsUser(long pk, long userPK) {
+	public boolean containsUser(long pk, long userPK) throws SystemException {
 		boolean finderClassNameCacheEnabled = OrganizationModelImpl.CACHE_ENABLED_USERS_ORGS;
 
 		String finderClassName = "Users_Orgs";
@@ -2174,13 +2178,18 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		}
 
 		if (result == null) {
-			Boolean value = Boolean.valueOf(containsUser.contains(pk, userPK));
+			try {
+				Boolean value = Boolean.valueOf(containsUser.contains(pk, userPK));
 
-			FinderCacheUtil.putResult(finderClassNameCacheEnabled,
-				finderClassName, finderMethodName, finderParams, finderArgs,
-				value);
+				FinderCacheUtil.putResult(finderClassNameCacheEnabled,
+					finderClassName, finderMethodName, finderParams,
+					finderArgs, value);
 
-			return value.booleanValue();
+				return value.booleanValue();
+			}
+			catch (Exception e) {
+				throw processException(e);
+			}
 		}
 		else {
 			return ((Boolean)result).booleanValue();
@@ -2200,8 +2209,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			addUser.add(pk, userPK);
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2213,8 +2222,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			addUser.add(pk, user.getPrimaryKey());
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2227,8 +2236,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addUser.add(pk, userPK);
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2242,8 +2251,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addUser.add(pk, user.getPrimaryKey());
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2254,8 +2263,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			clearUsers.clear(pk);
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2266,8 +2275,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			removeUser.remove(pk, userPK);
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2279,8 +2288,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 		try {
 			removeUser.remove(pk, user.getPrimaryKey());
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2293,8 +2302,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				removeUser.remove(pk, userPK);
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2308,8 +2317,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				removeUser.remove(pk, user.getPrimaryKey());
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2324,8 +2333,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addUser.add(pk, userPK);
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2341,8 +2350,8 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 				addUser.add(pk, user.getPrimaryKey());
 			}
 		}
-		catch (DataAccessException dae) {
-			throw new SystemException(dae);
+		catch (Exception e) {
+			throw processException(e);
 		}
 		finally {
 			FinderCacheUtil.clearCache("Users_Orgs");
@@ -2408,23 +2417,17 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 	protected ClearUsers clearUsers;
 	protected RemoveUser removeUser;
 
-	protected class ContainsGroup extends MappingSqlQuery {
+	protected class ContainsGroup {
 		protected ContainsGroup(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(), _SQL_CONTAINSGROUP);
+			super();
 
-			declareParameter(new SqlParameter(Types.BIGINT));
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
-		}
-
-		protected Object mapRow(ResultSet rs, int rowNumber)
-			throws SQLException {
-			return new Integer(rs.getInt("COUNT_VALUE"));
+			_mappingSqlQuery = MappingSqlQueryFactoryUtil.getMappingSqlQuery(getDataSource(),
+					_SQL_CONTAINSGROUP,
+					new int[] { Types.BIGINT, Types.BIGINT }, RowMapper.COUNT);
 		}
 
 		protected boolean contains(long organizationId, long groupId) {
-			List<Integer> results = execute(new Object[] {
+			List<Integer> results = _mappingSqlQuery.execute(new Object[] {
 						new Long(organizationId), new Long(groupId)
 					});
 
@@ -2438,79 +2441,71 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 
 			return false;
 		}
+
+		private MappingSqlQuery _mappingSqlQuery;
 	}
 
-	protected class AddGroup extends SqlUpdate {
+	protected class AddGroup {
 		protected AddGroup(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(),
-				"INSERT INTO Groups_Orgs (organizationId, groupId) VALUES (?, ?)");
-
+			_sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(getDataSource(),
+					"INSERT INTO Groups_Orgs (organizationId, groupId) VALUES (?, ?)",
+					new int[] { Types.BIGINT, Types.BIGINT });
 			_persistenceImpl = persistenceImpl;
-
-			declareParameter(new SqlParameter(Types.BIGINT));
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
 		}
 
 		protected void add(long organizationId, long groupId) {
 			if (!_persistenceImpl.containsGroup.contains(organizationId, groupId)) {
-				update(new Object[] { new Long(organizationId), new Long(
-							groupId) });
+				_sqlUpdate.update(new Object[] {
+						new Long(organizationId), new Long(groupId)
+					});
 			}
 		}
 
+		private SqlUpdate _sqlUpdate;
 		private OrganizationPersistenceImpl _persistenceImpl;
 	}
 
-	protected class ClearGroups extends SqlUpdate {
+	protected class ClearGroups {
 		protected ClearGroups(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(),
-				"DELETE FROM Groups_Orgs WHERE organizationId = ?");
-
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
+			_sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(getDataSource(),
+					"DELETE FROM Groups_Orgs WHERE organizationId = ?",
+					new int[] { Types.BIGINT });
 		}
 
 		protected void clear(long organizationId) {
-			update(new Object[] { new Long(organizationId) });
+			_sqlUpdate.update(new Object[] { new Long(organizationId) });
 		}
+
+		private SqlUpdate _sqlUpdate;
 	}
 
-	protected class RemoveGroup extends SqlUpdate {
+	protected class RemoveGroup {
 		protected RemoveGroup(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(),
-				"DELETE FROM Groups_Orgs WHERE organizationId = ? AND groupId = ?");
-
-			declareParameter(new SqlParameter(Types.BIGINT));
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
+			_sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(getDataSource(),
+					"DELETE FROM Groups_Orgs WHERE organizationId = ? AND groupId = ?",
+					new int[] { Types.BIGINT, Types.BIGINT });
 		}
 
 		protected void remove(long organizationId, long groupId) {
-			update(new Object[] { new Long(organizationId), new Long(groupId) });
+			_sqlUpdate.update(new Object[] {
+					new Long(organizationId), new Long(groupId)
+				});
 		}
+
+		private SqlUpdate _sqlUpdate;
 	}
 
-	protected class ContainsUser extends MappingSqlQuery {
+	protected class ContainsUser {
 		protected ContainsUser(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(), _SQL_CONTAINSUSER);
+			super();
 
-			declareParameter(new SqlParameter(Types.BIGINT));
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
-		}
-
-		protected Object mapRow(ResultSet rs, int rowNumber)
-			throws SQLException {
-			return new Integer(rs.getInt("COUNT_VALUE"));
+			_mappingSqlQuery = MappingSqlQueryFactoryUtil.getMappingSqlQuery(getDataSource(),
+					_SQL_CONTAINSUSER,
+					new int[] { Types.BIGINT, Types.BIGINT }, RowMapper.COUNT);
 		}
 
 		protected boolean contains(long organizationId, long userId) {
-			List<Integer> results = execute(new Object[] {
+			List<Integer> results = _mappingSqlQuery.execute(new Object[] {
 						new Long(organizationId), new Long(userId)
 					});
 
@@ -2524,59 +2519,58 @@ public class OrganizationPersistenceImpl extends BasePersistenceImpl
 
 			return false;
 		}
+
+		private MappingSqlQuery _mappingSqlQuery;
 	}
 
-	protected class AddUser extends SqlUpdate {
+	protected class AddUser {
 		protected AddUser(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(),
-				"INSERT INTO Users_Orgs (organizationId, userId) VALUES (?, ?)");
-
+			_sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(getDataSource(),
+					"INSERT INTO Users_Orgs (organizationId, userId) VALUES (?, ?)",
+					new int[] { Types.BIGINT, Types.BIGINT });
 			_persistenceImpl = persistenceImpl;
-
-			declareParameter(new SqlParameter(Types.BIGINT));
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
 		}
 
 		protected void add(long organizationId, long userId) {
 			if (!_persistenceImpl.containsUser.contains(organizationId, userId)) {
-				update(new Object[] { new Long(organizationId), new Long(userId) });
+				_sqlUpdate.update(new Object[] {
+						new Long(organizationId), new Long(userId)
+					});
 			}
 		}
 
+		private SqlUpdate _sqlUpdate;
 		private OrganizationPersistenceImpl _persistenceImpl;
 	}
 
-	protected class ClearUsers extends SqlUpdate {
+	protected class ClearUsers {
 		protected ClearUsers(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(),
-				"DELETE FROM Users_Orgs WHERE organizationId = ?");
-
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
+			_sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(getDataSource(),
+					"DELETE FROM Users_Orgs WHERE organizationId = ?",
+					new int[] { Types.BIGINT });
 		}
 
 		protected void clear(long organizationId) {
-			update(new Object[] { new Long(organizationId) });
+			_sqlUpdate.update(new Object[] { new Long(organizationId) });
 		}
+
+		private SqlUpdate _sqlUpdate;
 	}
 
-	protected class RemoveUser extends SqlUpdate {
+	protected class RemoveUser {
 		protected RemoveUser(OrganizationPersistenceImpl persistenceImpl) {
-			super(persistenceImpl.getDataSource(),
-				"DELETE FROM Users_Orgs WHERE organizationId = ? AND userId = ?");
-
-			declareParameter(new SqlParameter(Types.BIGINT));
-			declareParameter(new SqlParameter(Types.BIGINT));
-
-			compile();
+			_sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(getDataSource(),
+					"DELETE FROM Users_Orgs WHERE organizationId = ? AND userId = ?",
+					new int[] { Types.BIGINT, Types.BIGINT });
 		}
 
 		protected void remove(long organizationId, long userId) {
-			update(new Object[] { new Long(organizationId), new Long(userId) });
+			_sqlUpdate.update(new Object[] {
+					new Long(organizationId), new Long(userId)
+				});
 		}
+
+		private SqlUpdate _sqlUpdate;
 	}
 
 	private static final String _SQL_GETGROUPS = "SELECT {Group_.*} FROM Group_ INNER JOIN Groups_Orgs ON (Groups_Orgs.groupId = Group_.groupId) WHERE (Groups_Orgs.organizationId = ?)";
